+++ lib/lelylan/client.rb
@@ -9,6 +9,7 @@ require 'lelylan/client/type'
 require 'lelylan/client/property'
 require 'lelylan/client/function'
 require 'lelylan/client/status'
+require 'lelylan/client/category'
 require 'lelylan/client/location'
 require 'lelylan/client/subscription'
 require 'lelylan/client/physical'
@@ -37,6 +38,7 @@ module Lelylan
     include Lelylan::Client::Property
     include Lelylan::Client::Function
     include Lelylan::Client::Status
+    include Lelylan::Client::Category
     include Lelylan::Client::Location
     include Lelylan::Client::Subscription
     include Lelylan::Client::Physical